# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Access group governing which features are available.
 */
export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface Profile {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: UserRole;
  /** Whether the user may create or amend cheques (editor or admin). */
  canEdit: boolean;
  /** Whether the user may manage accounts and departments (admin). */
  canManage: boolean;
}

export interface Account {
  id: number;
  code: string;
  name: string;
  active: boolean;
}

export interface AccountInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  active?: boolean;
}

export interface Department {
  id: number;
  code: string;
  name: string;
  active: boolean;
}

export interface DepartmentInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  active?: boolean;
}

export type ChequeStatus = typeof ChequeStatus[keyof typeof ChequeStatus];


export const ChequeStatus = {
  outstanding: 'outstanding',
  cleared: 'cleared',
  cancelled: 'cancelled',
} as const;

export interface Cheque {
  id: number;
  chequeNumber: string;
  accountId: number;
  accountName: string;
  departmentId: number;
  departmentName: string;
  payee: string;
  /** Cheque amount in major currency units. */
  amount: number;
  issueDate: string;
  status: ChequeStatus;
  /** @nullable */
  clearedDate: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ChequeInput {
  /** @minLength 1 */
  chequeNumber: string;
  accountId: number;
  departmentId: number;
  /** @minLength 1 */
  payee: string;
  amount: number;
  issueDate: string;
  status?: ChequeStatus;
  /** @nullable */
  clearedDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ChequeUpdate {
  /** @minLength 1 */
  chequeNumber?: string;
  accountId?: number;
  departmentId?: number;
  /** @minLength 1 */
  payee?: string;
  amount?: number;
  issueDate?: string;
  status?: ChequeStatus;
  /** @nullable */
  clearedDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface AccountReportRow {
  accountId: number;
  accountCode: string;
  accountName: string;
  chequeCount: number;
  totalAmount: number;
  outstandingCount: number;
  outstandingAmount: number;
}

export interface DepartmentReportRow {
  departmentId: number;
  departmentCode: string;
  departmentName: string;
  chequeCount: number;
  totalAmount: number;
  outstandingCount: number;
  outstandingAmount: number;
}

export interface OutstandingReport {
  totalCount: number;
  totalAmount: number;
  cheques: Cheque[];
}

export interface DashboardSummary {
  totalCheques: number;
  totalAmount: number;
  outstandingCount: number;
  outstandingAmount: number;
  clearedCount: number;
  clearedAmount: number;
  cancelledCount: number;
  accountCount: number;
  departmentCount: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListChequesParams = {
/**
 * Inclusive lower bound on issue date (YYYY-MM-DD).
 */
startDate?: string;
/**
 * Inclusive upper bound on issue date (YYYY-MM-DD).
 */
endDate?: string;
accountId?: number;
departmentId?: number;
status?: ChequeStatus;
/**
 * Free-text search over cheque number and payee.
 */
search?: string;
};

export type GetAccountsReportParams = {
startDate?: string;
endDate?: string;
};

export type GetDepartmentsReportParams = {
startDate?: string;
endDate?: string;
};

export type GetOutstandingReportParams = {
startDate?: string;
endDate?: string;
};