# lib/api-zod/src/generated/types/cheque.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChequeStatus } from './chequeStatus';

export interface Cheque {
  id: number;
  chequeNumber: string;
  accountId: number;
  accountName: string;
  departmentId: number;
  departmentName: string;
  payee: string;
  /** Cheque amount in major currency units. */
  amount: number;
  issueDate: Date;
  status: ChequeStatus;
  /** @nullable */
  clearedDate: Date | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  updatedAt: Date;
}