# lib/api-zod/src/generated/types/chequeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChequeStatus } from './chequeStatus';

export interface ChequeInput {
  /** @minLength 1 */
  chequeNumber: string;
  accountId: number;
  departmentId: number;
  /** @minLength 1 */
  payee: string;
  amount: number;
  issueDate: Date;
  status?: ChequeStatus;
  /** @nullable */
  clearedDate?: Date | null;
  /** @nullable */
  notes?: string | null;
}